re-use `writeFileSafely` in other spots
