+++ packages/@tailwindcss-upgrade/src/codemods/template/migrate.ts
@@ -1,4 +1,3 @@
-import { randomUUID } from 'node:crypto'
 import fs from 'node:fs/promises'
 import path, { extname } from 'node:path'
 import {
@@ -10,6 +9,7 @@ import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
 import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
 import { DefaultMap } from '../../../../tailwindcss/src/utils/default-map'
 import { spliceChangesIntoString, type StringChange } from '../../utils/splice-changes-into-string'
+import { writeFileSafely } from '../../utils/write-file-safely'
 import { extractRawCandidates } from './candidates'
 import { isSafeMigration } from './is-safe-migration'
 import { migrateAutomaticVarInjection } from './migrate-automatic-var-injection'
@@ -129,46 +129,4 @@ export async function migrate(designSystem: DesignSystem, userConfig: Config | n
   if (migrated.trim() === '') return // Emptied out, something went horribly wrong
 
   await writeFileSafely(fullPath, migrated)
-}
-
-async function writeFileSafely(file: string, contents: string) {
-  // Start by creating a new file in the current directory that is guaranteed to
-  // be unique (via `uuid`). We can embed the `process.id` in case we need to
-  // debug things later.
-  //
-  // While we can write this to a more global `/tmp` folder, I want to be 100%
-  // sure that we are on the same file system (same drive) so the rename
-  // operation is atomic. Once the file is written, we will rename the file. If
-  // this fails, the old file is still intact, if it works we have an updated
-  // file.
-  //
-  // If this still causes problems (but it will slow things down):
-  // 1. We could make sure that we inherit the file permissions
-  // 2. Use an explicit fsync to force a flush to disk
-  let temporaryFile = path.join(
-    path.dirname(file),
-    `.${path.basename(file)}.tailwind-upgrade.${process.pid}.${randomUUID()}.tmp`,
-  )
-
-  // Write file uses the `w` flag by default, which is defined as:
-  // > Open file for writing. The file is created (if it does not exist) or truncated (if it exists).
-  // > https://nodejs.org/api/fs.html#file-system-flags
-  //
-  // Which means that if this function is actively running, and you cancel the
-  // process at the wrong time, then the truncated files are present. Since all
-  // these migrations happen in parallel, multiple files are open and available
-  // to be written to, it could mean in multiple truncated files.
-  //
-  // Writing to a temp file first means that if the process is cancelled at this
-  // point, that the old original file is still correct.
-  //
-  // The rename part should be atomic (especially because we guarantee it to be
-  // on the same file system) so this either succeeds or doesn't happen.
-  try {
-    await fs.writeFile(temporaryFile, contents, 'utf8')
-    await fs.rename(temporaryFile, file)
-  } catch (error) {
-    await fs.unlink(temporaryFile).catch(() => {})
-    throw error
-  }
 }
